Abstract code to DRY it up
